@@ -2,15 +2,17 @@
   "name": "full-contact-manager",
   "version": "0.1.0",
   "private": true,
+  "homepage": "contactApp",
   "dependencies": {
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.18.0",
     "react-scripts": "4.0.3",
-    "uuidv4": "^6.2.11",
+    "uuidv4": "^6.2.13",
+    "v4": "^0.0.1",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
