@@ -1,16 +1,16 @@
-import React, { } from 'react'
-import './contactList.css'
-import ContactCard from '../contactCard/ContactCard'
-import { Link } from 'react-router-dom';
+import React from 'react';
+import { NavLink } from 'react-router-dom';
+import ContactCard from '../contactCard/ContactCard';
+import './contactList.css';
 
 
 export default function ContactList(props) {
     // console.log(props);
-    //Remove contact
+
     const deleteContect = (id) => {
         props.getContactID(id);
     }
-    const contactlist = props.contactList.map((item, pos) => {
+    const contactList = props.contactList.length < 1 ? <h1>Nothing here</h1> : props.contactList.map(item => {
         return (
             <ContactCard data={item} key={item.id} clickDeleteContact={deleteContect} />
         )
@@ -18,13 +18,13 @@ export default function ContactList(props) {
 
     return (
         <div className="contactList">
-            <Link to='/add' style={{ textDecoration: 'none' }}>
+            <NavLink to='/add' style={{ textDecoration: 'none' }}>
                 <button style={{
                     padding: '5px 10px', backgroundColor: 'skyblue',
                     border: '1px solid black', display: 'block', marginLeft: '40%'
                 }}>Add Contact</button>
-            </Link>
-            {contactlist}
+            </NavLink>
+            {contactList}
         </div>
     );
 }
